Fixes citation and reference links
Breaks the update_references function into two functions, one to replace markdown links with
html links, and one to replace html links with markdown links. First, and html citation links
are converted to markdown, then the reference section is generated (with now-fixed backlinking
html), then citations are converted to html links, which are linked to by the backlinks in the
reference.